+++ src/components/Counters.vue
@@ -1,11 +1,11 @@
 <template>
     <div class="flex flex-wrap justify-center">
-      <p class="text-yellow-400 mx-2 text-2xl text-center mt-12"> 
+      <p class="text-yellow-400 mx-2 text-2xl text-center mt-12" v-if="!started"> 
             {{ currency }} :
             {{ gold }}
         </p>
 
-      <p class="text-red-400 mx-2 text-2xl text-center mt-12">
+      <p class="text-red-400 mx-2 text-2xl text-center mt-12" v-if="started">
           Damage :  
           {{ damage }}
         </p>
@@ -16,6 +16,9 @@
 <script>
     export default {
         computed: {
+            started(){
+                return this.$store.state.started;
+            },
             gold(){
                 return this.$store.state.gold;
             },
